Restructured unit tests from NN.cs into three files.
